Automated Item and Handle scripts
Creates the FlyDirRef transform, the Handle (and create the handle poses) when adding the item script. Also add a name to the itemId that is the name of the transform you add the item component. For the handle script, it create the handles poses if they don't have orientations and add them automatically (instead of pressing Update Orientations)