Fixed AssetBundleBuilder stripping custom shaders
There appears to be a bug in the current Asset Bundle Builder. It is currently stripping out all shaders that are referenced inside of a bundle group rather than only if they are in a shared bundle.

This resolves custom shader exports if the shaders have been declared as an Addressable Asset.